Leonid_Smirnov_todoapp/requirements.txt
asgiref==3.2.10
astroid==2.7.2
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.1
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.8.0
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
Markdown==3.3.4
mccabe==0.6.1
platformdirs==2.2.0
pycodestyle==2.7.0
PyJWT==2.1.0
pylint==2.10.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
wrapt==1.12.1
